project_102/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
reportlab==4.0.7
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26
pillow>=10.2.0
PyMySQL==1.1.1
cryptography==42.0.5
mysql-connector-python==8.3.0
